requirements.txt
asgiref==3.7.2
attrs==23.1.0
awsebcli==3.20.10
autopep8==1.6.0
beautifulsoup4==4.12.2
boto3==1.20.26
botocore==1.23.54
cachetools==5.3.3
cassandra-driver==3.29.1
cement==2.8.2
certifi==2023.5.7
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2024.1
crispy-bootstrap5==0.7
cryptography==41.0.2
daphne==4.0.0
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==0.5.0
Django==4.2
django-admin-sortable2==2.1.10
django-anymail==9.1
django-bootstrap4==23.2
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-dbbackup==4.0.2
django-dotenv==1.4.2
django-environ==0.9.0
django-import-export==3.2.0
django-jazzmin==2.6.0
django-recaptcha==3.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-six==1.0.5
django-static-fontawesome==5.14.0.0
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
environs==9.5.0
geoip2==4.6.0
geomet==0.2.1.post1
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gunicorn==21.2.0
httplib2==0.22.0
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.3
marshmallow==3.19.0
maxminddb==2.4.0
moviepy==1.0.3
msal==1.30.0
msgpack==1.0.7
numpy==1.26.0
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycodestyle==2.8.0
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
PyJWT==2.3.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.1.0
rsa==4.9
s3transfer==0.5.2
service-identity==23.1.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stripe==7.0.0
tqdm==4.66.1
toml==0.10.2
twilio==8.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.17
whitenoise==5.2.0
django-extensions
razorpay